adding token

+++ src/services/users.js
@@ -3,6 +3,7 @@ import axios from 'axios'
 const api = axios.create({
   baseURL: 'http://localhost:3030'
 })
+const token = localStorage.getItem('token')
 
 export const getUsers = async () => {
   const request = await api.get('/users')
@@ -10,25 +11,19 @@ export const getUsers = async () => {
 }
 
 export const createUser = async (user) => {
-  api.post('/users', user)
+  return await api.post('/users', user, { headers: { 'x-access-token': token } })
 }
 
 export const deleteUser = async (id) => {
-  api.delete(`/users/${id}`)
+  return await api.delete(`/users/${id}`, { headers: { 'x-access-token': token } })
 }
 
 export const editUser = async (user) => {
-  api.put(`/users`, user)
+  return await api.put(`/users`, user, { headers: { 'x-access-token': token } })
 }
 
 export const loginUser = async (user) => {
-  const token = localStorage.getItem('token')
-  const headers = {
-    "x-access-token": token
-  };
-  return await api.post(`/login`, user, {
-    headers: headers
-  })
+  return await api.post(`/login`, user, { headers: { 'x-access-token': token } })
 }
 
 export const registerUser = async (user) => {
